Extract shared methods into logs import service and lint
